+++ test/explain/test_hetero_explainer.py
@@ -8,7 +8,6 @@ from torch_geometric.explain import (
     HeteroExplanation,
 )
 from torch_geometric.explain.config import ExplanationType
-from torch_geometric.nn import SAGEConv, to_hetero
 
 
 def get_edge_index(num_src_nodes, num_dst_nodes, num_edges):
@@ -28,28 +27,13 @@ def data():
     return data
 
 
-class GraphSAGE(torch.nn.Module):
-    def __init__(self):
-        super().__init__()
-        self.conv1 = SAGEConv((-1, -1), 32)
-        self.conv2 = SAGEConv((-1, -1), 1)
-
-    def forward(self, x, edge_index):
-        x = self.conv1(x, edge_index).relu()
-        return self.conv2(x, edge_index)
-
-
-class HeteroSAGE(torch.nn.Module):
-    def __init__(self, metadata):
-        super().__init__()
-        self.graph_sage = to_hetero(GraphSAGE(), metadata, debug=False)
-
+class DummyModel(torch.nn.Module):
     def forward(self, x_dict, edge_index_dict, *args) -> torch.Tensor:
-        return self.graph_sage(x_dict, edge_index_dict)['paper']
+        return x_dict['paper'].mean().view(-1)
 
 
 def test_get_prediction(data):
-    model = HeteroSAGE(data.metadata())
+    model = DummyModel()
     assert model.training
 
     explainer = Explainer(
@@ -64,14 +48,13 @@ def test_get_prediction(data):
     )
     pred = explainer.get_prediction(data.x_dict, data.edge_index_dict)
     assert model.training
-
-    assert pred.size() == (8, 1)
+    assert pred.size() == (1, )
 
 
 @pytest.mark.parametrize('target', [None, torch.randn(2)])
 @pytest.mark.parametrize('explanation_type', [x for x in ExplanationType])
 def test_forward(data, target, explanation_type):
-    model = HeteroSAGE(data.metadata())
+    model = DummyModel()
 
     explainer = Explainer(
         model,
@@ -88,21 +71,25 @@ def test_forward(data, target, explanation_type):
         with pytest.raises(ValueError):
             explainer(data.x_dict, data.edge_index_dict, target=target)
     else:
-        explanation = explainer(data.x_dict, data.edge_index_dict,
-                                target=target)
+        explanation = explainer(
+            data.x_dict,
+            data.edge_index_dict,
+            target=target
+            if explanation_type == ExplanationType.phenomenon else None,
+        )
+        assert model.training
         assert isinstance(explanation, HeteroExplanation)
         assert 'node_feat_mask' in explanation.available_explanations
-        for node_type, node_feat_mask in explanation.node_feat_mask.items():
-            assert node_feat_mask.size() == data[node_type].x.size()
+        for key in explanation.node_types:
+            assert explanation[key].node_feat_mask.size() == data[key].x.size()
 
 
 @pytest.mark.parametrize('threshold_value', [0.2, 0.5, 0.8])
 @pytest.mark.parametrize('node_mask_type', ['object', 'attributes'])
 def test_hard_threshold(data, threshold_value, node_mask_type):
-    model = HeteroSAGE(data.metadata())
 
     explainer = Explainer(
-        model,
+        DummyModel(),
         algorithm=DummyExplainer(),
         explanation_type='model',
         node_mask_type=node_mask_type,
@@ -113,9 +100,7 @@ def test_hard_threshold(data, threshold_value, node_mask_type):
         ),
         threshold_config=('hard', threshold_value),
     )
-
     explanation = explainer(data.x_dict, data.edge_index_dict)
-
     if node_mask_type == 'object':
         assert 'node_mask' in explanation.available_explanations
     elif node_mask_type == 'attributes':
@@ -123,19 +108,16 @@ def test_hard_threshold(data, threshold_value, node_mask_type):
     assert 'edge_mask' in explanation.available_explanations
 
     for key in explanation.available_explanations:
-        masks = explanation[key]
-        for mask in masks.values():
+        for mask in explanation.collect(key).values():
             assert set(mask.unique().tolist()).issubset({0, 1})
 
 
-@pytest.mark.parametrize('threshold_value', list(range(1, 10)))
+@pytest.mark.parametrize('threshold_value', [1, 5, 10])
 @pytest.mark.parametrize('threshold_type', ['topk', 'topk_hard'])
 @pytest.mark.parametrize('node_mask_type', ['object', 'attributes'])
 def test_topk_threshold(data, threshold_value, threshold_type, node_mask_type):
-    model = HeteroSAGE(data.metadata())
-
     explainer = Explainer(
-        model,
+        DummyModel(),
         algorithm=DummyExplainer(),
         explanation_type='model',
         node_mask_type=node_mask_type,
@@ -155,8 +137,7 @@ def test_topk_threshold(data, threshold_value, threshold_type, node_mask_type):
     assert 'edge_mask' in explanation.available_explanations
 
     for key in explanation.available_explanations:
-        masks = explanation[key]
-        for mask in masks.values():
+        for mask in explanation.collect(key).values():
             if threshold_type == 'topk':
                 assert (mask > 0).sum() == min(mask.numel(), threshold_value)
                 assert ((mask == 0).sum() == mask.numel() -